mandelbrot scene added
Added a scene for calculating a 2D Mandelbrot fractal, and spent A LOT of time trying to learn the math behind it and messing around with different values.
i haven't figured out the zoom perfectly yet, so that's what I'd like to work on next, but we'll see. I like the result so far

also scenes handle their own keyboard input, instead of the rendering handling things like camera movement.